requirements.txt
streamlit>=1.36
python-dateutil>=2.9
# Gemini SDK (new). If this fails in your env, see app.py fallback notes.
google-genai>=0.6.0
google-generativeai>=0.8.0
